app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
redis
langchain
anthropic
pydantic
python-dotenv
httpx 